remove constants.js
