+++ lib/eventsource/eventsource-stream.js
@@ -1,14 +1,34 @@
 'use strict'
 const { Transform } = require('node:stream')
-const {
-  BOM,
-  CR,
-  LF,
-  COLON,
-  SPACE,
-  validMessageEventFieldNames
-} = require('./constants')
 const { isASCIINumber, isValidLastEventId } = require('./util')
+
+/**
+ * @type {number[]} BOM
+ */
+const BOM = [0xEF, 0xBB, 0xBF]
+/**
+ * @type {10} LF
+ */
+const LF = 0x0A
+/**
+ * @type {13} CR
+ */
+const CR = 0x0D
+/**
+ * @type {58} COLON
+ */
+const COLON = 0x3A
+/**
+ * @type {32} SPACE
+ */
+const SPACE = 0x20
+
+const validMessageEventFieldNames = [
+  'data',
+  'event',
+  'id',
+  'retry'
+]
 
 /**
  * @typedef {object} EventSourceStreamEvent
