+++ lib/eventsource/eventsource.js
@@ -7,7 +7,6 @@ const { HeadersList } = require('../fetch/headers')
 const { makeRequest } = require('../fetch/request')
 const { getGlobalOrigin } = require('../fetch/global')
 const { webidl } = require('../fetch/webidl')
-const { CONNECTING, OPEN, CLOSED, defaultReconnectionTime, ANONYMOUS, USE_CREDENTIALS } = require('./constants')
 const { EventSourceStream } = require('./eventsource-stream')
 const { parseMIMEType } = require('../fetch/dataURL')
 const { MessageEvent, ErrorEvent } = require('../websocket/events')
@@ -15,6 +14,57 @@ const { isNetworkError } = require('../fetch/response')
 const { getGlobalDispatcher } = require('../global')
 
 let experimentalWarned = false
+
+/**
+ * A reconnection time, in milliseconds. This must initially be an implementation-defined value,
+ * probably in the region of a few seconds.
+ *
+ * In Comparison:
+ * - Chrome uses 3000ms.
+ * - Deno uses 5000ms.
+ *
+ * @type {3000}
+ */
+const defaultReconnectionTime = 3000
+
+/**
+ * The readyState attribute represents the state of the connection.
+ * @enum
+ * @readonly
+ * @see https://html.spec.whatwg.org/multipage/server-sent-events.html#dom-eventsource-readystate-dev
+ */
+
+/**
+ * The connection has not yet been established, or it was closed and the user
+ * agent is reconnecting.
+ * @type {0}
+ */
+const CONNECTING = 0
+
+/**
+ * The user agent has an open connection and is dispatching events as it
+ * receives them.
+ * @type {1}
+ */
+const OPEN = 1
+
+/**
+ * The connection is not open, and the user agent is not trying to reconnect.
+ * @type {2}
+ */
+const CLOSED = 2
+
+/**
+ * Requests for the element will have their mode set to "cors" and their credentials mode set to "same-origin".
+ * @type {'anonymous'}
+ */
+const ANONYMOUS = 'anonymous'
+
+/**
+ * Requests for the element will have their mode set to "cors" and their credentials mode set to "include".
+ * @type {'use-credentials'}
+ */
+const USE_CREDENTIALS = 'use-credentials'
 
 /**
  * @typedef {object} EventSourceInit
@@ -438,5 +488,6 @@ webidl.converters.EventSourceInitDict = webidl.dictionaryConverter([
 ])
 
 module.exports = {
-  EventSource
+  EventSource,
+  defaultReconnectionTime
 }
