+++ test/eventsource/eventsource-reconnect.js
@@ -4,8 +4,7 @@ const assert = require('node:assert')
 const events = require('node:events')
 const http = require('node:http')
 const { test, describe } = require('node:test')
-const { defaultReconnectionTime } = require('../../lib/eventsource/constants')
-const { EventSource } = require('../../lib/eventsource/eventsource')
+const { EventSource, defaultReconnectionTime } = require('../../lib/eventsource/eventsource')
 
 describe('EventSource - reconnect', () => {
   test('Should reconnect on connection close', async () => {
